cognitive_agent_openclaw/requirements.txt
google-genai>=1.0.0
paho-mqtt>=1.6.1,<2.0
requests>=2.28.0
